Support Scala 3 & Play 3 - drop some old versions
Play v2.9 introduces support for Scala 3 - Play v3.0 also switches from
Akka to Pekko, which is great from a licensing perspective (Play v2.9
had to ship with old versions of Akka to avoid the BSL license:
https://www.playframework.com/documentation/3.0.x/General#How-Play-Deals-with-Akkas-License-Change ).

Here, we're upgrading `play-secret-rotation` to support these new latest
versions of Play, and add the new welcome Scala 3 support! To reduce
maintenance build complexity, we're also dropping support for Play 2.6,
and Scala 2.12.

Note, Matthias Kurz made some interesting suggestions about
restructuring the multi-Play support in this project and also
`google-play-auth`:

https://github.com/guardian/frontend/pull/26677#issuecomment-1801443253

...however, for `play-secret-rotation` at least, the changes didn't
appear necessary, as `play-secret-rotation` actually has a fairly
minimal dependency on Play, and doesn't touch Akka/Pekko related code.
