+++ play/play-v30/RotatingSecretComponents.scala
@@ -1,0 +1,1 @@
+../play-v26/RotatingSecretComponents.scala
